Read serial function updated to display characters as they come in a line and other minor adjustments

+++ lab3_test_code/src/serial.c
@@ -28,7 +28,7 @@
 
 int main(int argc, char* argv[])
 {
-    uint32_t data;;
+    uint16_t data;;
     if (argc == 3)
     {
         serialOpen();
